hopefully fix executables
