+++ randovania/server/socketio.py
@@ -2,23 +2,22 @@ from __future__ import annotations
 
 import typing
 from contextlib import asynccontextmanager
-from typing import TYPE_CHECKING, Any, Concatenate
+from typing import TYPE_CHECKING, Concatenate
 
 from prometheus_client import Gauge
-from socketio import AsyncClient
 from socketio.exceptions import ConnectionRefusedError
 from socketio_handler import BaseSocketHandler, SocketManager, register_handler
 
 import randovania
+from randovania.network_common.signals import SioDataType, _args_to_sio_data
 from randovania.server import client_check
 
 if TYPE_CHECKING:
-    from collections.abc import Callable, Mapping, Sequence
+    from collections.abc import Callable
 
     from randovania.network_client.network_client import NetworkClient
     from randovania.server.server_app import AsyncCallable, Lifespan, RdvFastAPI, ServerApp
 
-type SioDataType = str | bytes | Mapping[str, SioDataType] | Sequence[SioDataType]
 EventHandlerReturnType = SioDataType | tuple[SioDataType, ...] | None
 
 
@@ -86,13 +85,6 @@ def get_socket_handler(sa: ServerApp) -> type[BaseSocketHandler]:
     return RdvSocketHandler
 
 
-def _args_to_sio_data(*args: Any) -> SioDataType | tuple[SioDataType, ...]:
-    if len(args) == 1:
-        return typing.cast("SioDataType", args[0])
-    else:
-        return typing.cast("tuple[SioDataType, ...]", args)
-
-
 class ServerEventHandler[**P, RetT]:
     def __init__(self, fn: AsyncCallable[Concatenate[ServerApp, str, P], RetT], message: str):
         self.fn = fn
@@ -146,70 +138,5 @@ def server_event_handler[**P, RetT](
 
     def decorator(fn: AsyncCallable[Concatenate[ServerApp, str, P], RetT]) -> ServerEventHandler[P, RetT]:
         return ServerEventHandler(fn, message)
-
-    return decorator
-
-
-class ClientSignal[**P]:
-    def __init__(self, fn: AsyncCallable[P, None], message: str):
-        self.fn = fn
-        self.message = message
-
-    async def __call__(self, *args: P.args, **kwargs: P.kwargs) -> None:
-        return await self.fn(*args, **kwargs)
-
-    def emit(
-        self,
-        sa: ServerApp,
-        to: str | None = None,
-        room: str | None = None,
-        namespace: str | None = None,
-    ) -> AsyncCallable[P, None]:
-        """
-        Returns an async callable which, when called and awaited, uses the `ServerApp`
-        to emit this signal. Provides full typing support, so it's preferred over directly
-        calling `sa.sio.emit()`.
-        """
-
-        async def inner(*args: P.args, **kwargs: P.kwargs) -> None:
-            await sa.sio.emit(self.message, _args_to_sio_data(*args), to=to, room=room, namespace=namespace)
-
-        return inner
-
-    def register(self, sio: AsyncClient, callback: AsyncCallable[P, None]) -> None:
-        """
-        Registers the given callback with the SIO client on this signal's message.
-
-        Using this function allows checking that the signature of the registered callback
-        is compatible with this signal's expected signature.
-        """
-        sio.on(self.message, callback)
-
-
-def client_signal[**P](message: str) -> Callable[[AsyncCallable[P, None]], ClientSignal[P]]:
-    """
-    Transforms the callable into a `ClientSignal` for fully type-checked signal emission from the server.
-
-    Example usage::
-
-        @client_signal("multiplayer_binary_inventory")
-        async def WORLD_BINARY_INVENTORY(entry_id: str, user_id: int, raw_inventory: bytes) -> None: ...
-
-        class NetworkClient:
-            def __init__(self, sio: AsyncClient):
-                self.sio = sio
-
-                WORLD_BINARY_INVENTORY.register(self.sio, self._on_world_user_inventory_raw)
-
-            def _on_world_user_inventory_raw(self, entry_id: str, user_id: int, raw_inventory: bytes) -> None:
-                print(entry_id, user_id, raw_inventory)
-
-        # prints "'entry', 1234, b'4321'"
-        await WORLD_BINARY_INVENTORY.emit(ServerApp())("entry", 1234, b"4321")
-
-    """
-
-    def decorator(fn: AsyncCallable[P, None]) -> ClientSignal[P]:
-        return ClientSignal(fn, message)
 
     return decorator
